requirements-no-cuda.txt
numpy==1.22.2
psycopg2==2.9.3
requests==2.27.1
websockets==10.1
websocket-client==1.3.1